refactor: move provider to providers dir
